refactor: 패키지명 수정 및 크리에이터 변수명 수정
creator repository 패키지명 수정
creator creatorName -> name
