+++ src/main/java/apptive/london/bridge/domain/user/entity/Creator.java
@@ -15,27 +15,30 @@ import java.util.List;
 @EqualsAndHashCode(callSuper = true)
 @SuperBuilder
 public class Creator extends User {
-    private String creatorName;
+    private String name;
     private String gender;
+
 
     @ElementCollection
     @Builder.Default
     private List<String> channelLinks = new ArrayList<>();
 
     private String businessEmail;
+    private Boolean callStatus;
 
     public Creator() {
         super();
     }
 
-    public static Creator fromUser(User user, String creatorName, String gender, List<String> channelLinks, String businessEmail) {
+    public static Creator fromUser(User user, String name, String gender, List<String> channelLinks, String businessEmail, Long followerCount) {
         return Creator.builder()
                 .email(user.getEmail())
                 .password(user.getPassword())
                 .nickname(user.getNickname())
                 .birthday(user.getBirthday())
+                .createdDate(user.getCreatedDate())
                 .role(Role.CREATOR)
-                .creatorName(creatorName)
+                .name(name)
                 .gender(gender)
                 .channelLinks(channelLinks)
                 .businessEmail(businessEmail)
