+++ src/main/java/apptive/london/bridge/domain/user/repository/CreatorRepository.java
@@ -1,4 +1,4 @@
-package apptive.london.bridge.domain.user.repositoy;
+package apptive.london.bridge.domain.user.repository;
 
 import apptive.london.bridge.domain.user.entity.Creator;
 import org.springframework.data.jpa.repository.JpaRepository;
